Mayavi: show box

+++ src/python/espressomd/visualization.py
@@ -17,6 +17,7 @@ def mayavi_render(queue):
 	points.glyph.color_mode = 'color_by_scalar'
 	points.glyph.glyph_source.glyph_source.center = [0, 0, 0]
 	arrows = mlab.quiver3d([0],[0],[0], [0],[0],[0])
+	box = mlab.outline(extent=(0,0,0,0,0,0), color=(1,1,1))
 
 	@mlab.show
 	@mlab.animate(delay=10, ui=True)
@@ -24,15 +25,12 @@ def mayavi_render(queue):
 		f = mlab.gcf()
 		visual.set_viewer(f)
 
-		box = visual.box(x=0, y=0, z=0, length=0, height=0, width=0, color=visual.color.red)
-
 		running = False
 		while True:
 			coords, types, radii, N_changed, bonds, Nbonds_changed, boxl, box_changed = queue.get()
 
 			if box_changed or not running:
-				box.x, box.y, box.z = boxl/2
-				box.length, box.height, box.width = boxl
+				box.set(bounds=(0,boxl[0], 0,boxl[1], 0,boxl[2]))
 
 			if not N_changed:
 				points.mlab_source.set(x=coords[:,0], y=coords[:,1], z=coords[:,2], u=radii, v=radii, w=radii, scalars=types)
@@ -47,26 +45,32 @@ def mayavi_render(queue):
 			else:
 				arrows.mlab_source.reset(x=bonds[:,0], y=bonds[:,1], z=bonds[:,2], u=bonds[:,3], v=bonds[:,4], w=bonds[:,5])
 
-			print boxl
-
 			yield
 	animate()
+	raise Exception("Animation window closed")
 
 class mayavi_live:
 	def __init__(self, system):
 		self.system = system
 		self.queue = Queue()
 		self.process = Process(target=mayavi_render, args=(self.queue,))
+		self.process.daemon = True
 		self.process.start()
 		self.last_N = 1
 		self.last_Nbonds = 1
 		self.last_boxl = [0,0,0]
-		atexit.register(self.process.terminate)
+		atexit.register(self.stop)
+
+	def stop(self):
+		if self.process.is_alive():
+			self.process.terminate()
 
 	def __del__(self):
-		self.process.terminate()
+		self.stop()
 
 	def update(self):
+		if not self.process.is_alive():
+			raise Exception("Animation process terminated")
 		N = self.system.n_part
 		coords = numpy.empty((N,3))
 		types = numpy.empty(N, dtype=int)
@@ -94,7 +98,6 @@ class mayavi_live:
 			bond_coords[n+1,3:] = self.system.part[j].pos - self.system.part[i].pos
 
 		boxl = self.system.box_l
-		print boxl, type(boxl)
 
 		self.queue.put(( coords, types, radii, (self.last_N != N), 
 		                 bond_coords, (self.last_Nbonds != Nbonds),
@@ -103,6 +106,5 @@ class mayavi_live:
 		self.last_Nbonds = Nbonds
 		self.last_boxl = boxl
 
-# TODO: simulation box should be line-only
 # TODO: constraints
 # TODO CHECK: bonds
